refactor(converters): move converters in own files
