+++ api/controllers/v1/cave/create.js
@@ -7,8 +7,8 @@
 const CaveService = require('../../../services/CaveService');
 const ControllerService = require('../../../services/ControllerService');
 const ErrorService = require('../../../services/ErrorService');
-const MappingService = require('../../../services/mapping/MappingService');
 const RightService = require('../../../services/RightService');
+const { toCave } = require('../../../services/mapping/converters');
 
 const { checkRight } = sails.helpers;
 
@@ -96,7 +96,7 @@ module.exports = async (req, res) => {
       populatedCave,
       params,
       res,
-      MappingService.convertToCaveModel
+      toCave
     );
   } catch (e) {
     return ErrorService.getDefaultErrorHandler(res)(e);
