+++ api/controllers/v1/caver/update.js
@@ -1,7 +1,7 @@
 const ControllerService = require('../../../services/ControllerService');
 const ErrorService = require('../../../services/ErrorService');
-const MappingService = require('../../../services/mapping/MappingService');
 const RightService = require('../../../services/RightService');
+const { toCaver } = require('../../../services/mapping/converters');
 
 module.exports = async (req, res) => {
   // list of propreties to update
@@ -97,7 +97,7 @@ module.exports = async (req, res) => {
       updatedCaver,
       params,
       res,
-      MappingService.convertToCaverModel
+      toCaver
     );
   } catch (e) {
     return ErrorService.getDefaultErrorHandler(res)(e);
