+++ api/controllers/v1/comment/update.js
@@ -1,12 +1,12 @@
 const ControllerService = require('../../../services/ControllerService');
 const ErrorService = require('../../../services/ErrorService');
-const MappingService = require('../../../services/mapping/MappingService');
 const {
   NOTIFICATION_TYPES,
   NOTIFICATION_ENTITIES,
 } = require('../../../services/NotificationService');
 const NotificationService = require('../../../services/NotificationService');
 const RightService = require('../../../services/RightService');
+const { toComment } = require('../../../services/mapping/converters');
 
 const { checkRight } = sails.helpers;
 
@@ -94,7 +94,7 @@ module.exports = async (req, res) => {
       populatedComment,
       params,
       res,
-      MappingService.convertToCommentModel
+      toComment
     );
   } catch (e) {
     return ErrorService.getDefaultErrorHandler(res)(e);
