+++ api/controllers/v1/description/create.js
@@ -1,12 +1,12 @@
 const ControllerService = require('../../../services/ControllerService');
 const ErrorService = require('../../../services/ErrorService');
-const MappingService = require('../../../services/mapping/MappingService');
 const NotificationService = require('../../../services/NotificationService');
 const RightService = require('../../../services/RightService');
 const {
   NOTIFICATION_TYPES,
   NOTIFICATION_ENTITIES,
 } = require('../../../services/NotificationService');
+const { toDescription } = require('../../../services/mapping/converters');
 
 module.exports = async (req, res) => {
   // Check right
@@ -135,7 +135,7 @@ module.exports = async (req, res) => {
       newDescriptionPopulated,
       params,
       res,
-      MappingService.convertToDescriptionModel
+      toDescription
     );
   } catch (e) {
     return ErrorService.getDefaultErrorHandler(res)(e);
