+++ api/controllers/v1/document/find-by-caver-id.js
@@ -2,12 +2,8 @@ const ControllerService = require('../../../services/ControllerService');
 const DescriptionService = require('../../../services/DescriptionService');
 const DocumentService = require('../../../services/DocumentService');
 const ErrorService = require('../../../services/ErrorService');
-const {
-  convertToDocumentModel,
-} = require('../../../services/mapping/MappingService');
-const {
-  convertToListFromController,
-} = require('../../../services/mapping/utils');
+const { toDocument } = require('../../../services/mapping/converters');
+const { toListFromController } = require('../../../services/mapping/utils');
 
 module.exports = async (req, res) => {
   const caverId = req.param('caverId');
@@ -88,12 +84,7 @@ module.exports = async (req, res) => {
           documents,
           params,
           res,
-          (data) =>
-            convertToListFromController(
-              'documents',
-              data,
-              convertToDocumentModel
-            )
+          (data) => toListFromController('documents', data, toDocument)
         );
       } catch (e) {
         return ErrorService.getDefaultErrorHandler(res)(e);
