+++ api/controllers/v1/entrance-duplicate/find.js
@@ -1,6 +1,6 @@
 const ControllerService = require('../../../services/ControllerService');
 const EntranceService = require('../../../services/EntranceService');
-const MappingService = require('../../../services/mapping/MappingService');
+const { toEntranceDuplicate } = require('../../../services/mapping/converters');
 
 module.exports = async (req, res) => {
   if (
@@ -58,6 +58,6 @@ module.exports = async (req, res) => {
     duplicateFound,
     params,
     res,
-    MappingService.convertToEntranceDuplicateModel
+    toEntranceDuplicate
   );
 };
