+++ api/controllers/v1/entrance/update.js
@@ -1,7 +1,6 @@
 const ControllerService = require('../../../services/ControllerService');
 const EntranceService = require('../../../services/EntranceService');
 const ErrorService = require('../../../services/ErrorService');
-const MappingService = require('../../../services/mapping/MappingService');
 const NameService = require('../../../services/NameService');
 const {
   NOTIFICATION_TYPES,
@@ -9,6 +8,7 @@ const {
 } = require('../../../services/NotificationService');
 const NotificationService = require('../../../services/NotificationService');
 const RightService = require('../../../services/RightService');
+const { toEntrance } = require('../../../services/mapping/converters');
 
 const checkRight = sails.helpers.checkRight.with;
 
@@ -104,7 +104,7 @@ module.exports = async (req, res) => {
       updatedEntrance,
       params,
       res,
-      MappingService.convertToEntranceModel
+      toEntrance
     );
   } catch (e) {
     return ErrorService.getDefaultErrorHandler(res)(e);
