+++ api/controllers/v1/history/create.js
@@ -1,12 +1,12 @@
 const ControllerService = require('../../../services/ControllerService');
 const ErrorService = require('../../../services/ErrorService');
-const MappingService = require('../../../services/mapping/MappingService');
 const {
   NOTIFICATION_TYPES,
   NOTIFICATION_ENTITIES,
 } = require('../../../services/NotificationService');
 const NotificationService = require('../../../services/NotificationService');
 const RightService = require('../../../services/RightService');
+const { toHistory } = require('../../../services/mapping/converters');
 
 module.exports = async (req, res) => {
   // Check right
@@ -89,7 +89,7 @@ module.exports = async (req, res) => {
       newHistoryPopulated,
       params,
       res,
-      MappingService.convertToHistoryModel
+      toHistory
     );
   } catch (e) {
     return ErrorService.getDefaultErrorHandler(res)(e);
