+++ api/controllers/v1/notification/find-all.js
@@ -1,11 +1,7 @@
 const ControllerService = require('../../../services/ControllerService');
 const ErrorService = require('../../../services/ErrorService');
-const {
-  convertToNotificationModel,
-} = require('../../../services/mapping/MappingService');
-const {
-  convertToListFromController,
-} = require('../../../services/mapping/utils');
+const { toNotification } = require('../../../services/mapping/converters');
+const { toListFromController } = require('../../../services/mapping/utils');
 const NotificationService = require('../../../services/NotificationService');
 
 const MAX_SIZE = 50;
@@ -69,12 +65,7 @@ module.exports = async (req, res) => {
       populatedNotifications,
       params,
       res,
-      (data) =>
-        convertToListFromController(
-          'notifications',
-          data,
-          convertToNotificationModel
-        )
+      (data) => toListFromController('notifications', data, toNotification)
     );
   } catch (e) {
     return ErrorService.getDefaultErrorHandler(res)(e);
