+++ api/controllers/v1/rigging/create.js
@@ -1,6 +1,6 @@
 const ControllerService = require('../../../services/ControllerService');
 const ErrorService = require('../../../services/ErrorService');
-const MappingService = require('../../../services/mapping/MappingService');
+
 const {
   NOTIFICATION_TYPES,
   NOTIFICATION_ENTITIES,
@@ -8,6 +8,7 @@ const {
 const NotificationService = require('../../../services/NotificationService');
 const RightService = require('../../../services/RightService');
 const RiggingService = require('../../../services/RiggingService');
+const { toRigging } = require('../../../services/mapping/converters');
 
 module.exports = async (req, res) => {
   // Check right
@@ -128,7 +129,7 @@ module.exports = async (req, res) => {
       newRiggingPopulated,
       params,
       res,
-      MappingService.convertToRiggingModel
+      toRigging
     );
   } catch (e) {
     return ErrorService.getDefaultErrorHandler(res)(e);
